feat: consolidate d1 adapters
