+++ packages/adapter-d1/package.json
@@ -24,7 +24,8 @@
   },
   "scripts": {
     "dev": "DEV=true tsx helpers/build.ts",
-    "build": "tsx helpers/build.ts"
+    "build": "tsx helpers/build.ts",
+    "test": "vitest run"
   },
   "files": [
     "dist",
@@ -42,5 +43,8 @@
     "@cloudflare/workers-types": "4.20250214.0",
     "@prisma/driver-adapter-utils": "workspace:*",
     "ky": "1.7.5"
+  },
+  "devDependencies": {
+    "vitest": "3.0.9"
   }
 }
