+++ packages/adapter-d1/src/d1.ts
@@ -1,211 +1,35 @@
-/* eslint-disable @typescript-eslint/require-await */
-
-import type { D1Database, D1Response } from '@cloudflare/workers-types'
-import {
-  ConnectionInfo,
-  Debug,
-  DriverAdapterError,
-  IsolationLevel,
-  SqlDriverAdapter,
-  SqlDriverAdapterFactory,
-  SqlQuery,
-  SqlQueryable,
-  SqlResultSet,
-  Transaction,
-  TransactionOptions,
-} from '@prisma/driver-adapter-utils'
-import { blue, cyan, red, yellow } from 'kleur/colors'
+import { D1Database } from '@cloudflare/workers-types'
+import { SqlDriverAdapter, SqlDriverAdapterFactory } from '@prisma/driver-adapter-utils'
 
 import { name as packageName } from '../package.json'
-import { MAX_BIND_VALUES } from './constants'
-import { getColumnTypes, mapRow } from './conversion'
-import { convertDriverError } from './errors'
-import { cleanArg } from './utils'
+import { D1HTTPParams, isD1HTTPParams, PrismaD1HTTPAdapterFactory } from './d1-http'
+import { PrismaD1WorkerAdapterFactory } from './d1-worker'
 
-const debug = Debug('prisma:driver-adapter:d1')
-
-type D1ResultsWithColumnNames = [string[], unknown[][]]
-type PerformIOResult = D1ResultsWithColumnNames | D1Response
-type StdClient = D1Database
-
-/**
- * Env binding for Cloudflare D1.
- */
-class D1Queryable<ClientT extends StdClient> implements SqlQueryable {
+// This is a wrapper type that can conform to either `SqlDriverAdapterFactory` or
+// `SqlMigrationAwareDriverAdapterFactory`, depending on the type of the argument passed to the
+// constructor. If the argument is of type `D1HTTPParams`, it will leverage
+// `PrismaD1HTTPAdapterFactory`. Otherwise, it will use `PrismaD1WorkerAdapterFactory`.
+export class PrismaD1<Args extends D1Database | D1HTTPParams = D1Database> implements PrismaD1Interface<Args> {
   readonly provider = 'sqlite'
   readonly adapterName = packageName
 
-  constructor(protected readonly client: ClientT) {}
+  connect!: PrismaD1Interface<Args>['connect']
+  connectToShadowDb!: PrismaD1Interface<Args>['connectToShadowDb']
 
-  /**
-   * Execute a query given as SQL, interpolating the given parameters.
-   */
-  async queryRaw(query: SqlQuery): Promise<SqlResultSet> {
-    const tag = '[js::query_raw]'
-    debug(`${tag} %O`, query)
-
-    const data = await this.performIO(query)
-    const convertedData = this.convertData(data as D1ResultsWithColumnNames)
-    return convertedData
-  }
-
-  private convertData(ioResult: D1ResultsWithColumnNames): SqlResultSet {
-    const columnNames = ioResult[0]
-    const results = ioResult[1]
-
-    if (results.length === 0) {
-      return {
-        columnNames: [],
-        columnTypes: [],
-        rows: [],
-      }
-    }
-
-    const columnTypes = Object.values(getColumnTypes(columnNames, results))
-    const rows = results.map((value) => mapRow(value, columnTypes))
-
-    return {
-      columnNames,
-      // * Note: without Object.values the array looks like
-      // * columnTypes: [ id: 128 ],
-      // * and errors with:
-      // * ✘ [ERROR] A hanging Promise was canceled. This happens when the worker runtime is waiting for a Promise from JavaScript to resolve, but has detected that the Promise cannot possibly ever resolve because all code and events related to the Promise's I/O context have already finished.
-      columnTypes,
-      rows,
-    }
-  }
-
-  /**
-   * Execute a query given as SQL, interpolating the given parameters and
-   * returning the number of affected rows.
-   * Note: Queryable expects a u64, but napi.rs only supports u32.
-   */
-  async executeRaw(query: SqlQuery): Promise<number> {
-    const tag = '[js::execute_raw]'
-    debug(`${tag} %O`, query)
-
-    const result = await this.performIO(query, true)
-    return (result as D1Response).meta.changes ?? 0
-  }
-
-  private async performIO(query: SqlQuery, executeRaw = false): Promise<PerformIOResult> {
-    try {
-      query.args = query.args.map((arg, i) => cleanArg(arg, query.argTypes[i]))
-
-      const stmt = this.client.prepare(query.sql).bind(...query.args)
-
-      if (executeRaw) {
-        return await stmt.run()
-      } else {
-        const [columnNames, ...rows] = await stmt.raw({ columnNames: true })
-        return [columnNames, rows]
-      }
-    } catch (e) {
-      onError(e as Error)
+  constructor(params: Args) {
+    if (isD1HTTPParams(params)) {
+      const factory = new PrismaD1HTTPAdapterFactory(params)
+      const self = this as PrismaD1Interface<D1HTTPParams>
+      self.connect = factory.connect.bind(factory)
+      self.connectToShadowDb = factory.connectToShadowDb.bind(factory)
+    } else {
+      const factory = new PrismaD1WorkerAdapterFactory(params)
+      const self = this as PrismaD1Interface<D1Database>
+      self.connect = factory.connect.bind(factory)
     }
   }
 }
 
-class D1Transaction extends D1Queryable<StdClient> implements Transaction {
-  constructor(client: StdClient, readonly options: TransactionOptions) {
-    super(client)
-  }
-
-  async commit(): Promise<void> {
-    debug(`[js::commit]`)
-  }
-
-  async rollback(): Promise<void> {
-    debug(`[js::rollback]`)
-  }
-}
-
-export class PrismaD1Adapter extends D1Queryable<StdClient> implements SqlDriverAdapter {
-  readonly tags = {
-    error: red('prisma:error'),
-    warn: yellow('prisma:warn'),
-    info: cyan('prisma:info'),
-    query: blue('prisma:query'),
-  }
-
-  alreadyWarned = new Set()
-
-  constructor(client: StdClient, private readonly release?: () => Promise<void>) {
-    super(client)
-  }
-
-  /**
-   * This will warn once per transaction
-   * e.g. the following two explicit transactions
-   * will only trigger _two_ warnings
-   *
-   * ```ts
-   * await prisma.$transaction([ ...queries ])
-   * await prisma.$transaction([ ...moreQueries ])
-   * ```
-   */
-  private warnOnce = (key: string, message: string, ...args: unknown[]) => {
-    if (!this.alreadyWarned.has(key)) {
-      this.alreadyWarned.add(key)
-      console.info(`${this.tags.warn} ${message}`, ...args)
-    }
-  }
-
-  async executeScript(script: string): Promise<void> {
-    try {
-      await this.client.exec(script)
-    } catch (error) {
-      onError(error as Error)
-    }
-  }
-
-  getConnectionInfo(): ConnectionInfo {
-    return {
-      maxBindValues: MAX_BIND_VALUES,
-      supportsRelationJoins: false,
-    }
-  }
-
-  async startTransaction(isolationLevel?: IsolationLevel): Promise<Transaction> {
-    if (isolationLevel && isolationLevel !== 'SERIALIZABLE') {
-      throw new DriverAdapterError({
-        kind: 'InvalidIsolationLevel',
-        level: isolationLevel,
-      })
-    }
-
-    this.warnOnce(
-      'D1 Transaction',
-      "Cloudflare D1 does not support transactions yet. When using Prisma's D1 adapter, implicit & explicit transactions will be ignored and run as individual queries, which breaks the guarantees of the ACID properties of transactions. For more details see https://pris.ly/d/d1-transactions",
-    )
-
-    const options: TransactionOptions = {
-      usePhantomQuery: true,
-    }
-
-    const tag = '[js::startTransaction]'
-    debug('%s options: %O', tag, options)
-
-    return new D1Transaction(this.client, options)
-  }
-
-  async dispose(): Promise<void> {
-    await this.release?.()
-  }
-}
-
-export class PrismaD1AdapterFactory implements SqlDriverAdapterFactory {
-  readonly provider = 'sqlite'
-  readonly adapterName = packageName
-
-  constructor(private client: StdClient) {}
-
-  async connect(): Promise<SqlDriverAdapter> {
-    return new PrismaD1Adapter(this.client, async () => {})
-  }
-}
-
-function onError(error: Error): never {
-  console.error('Error in performIO: %O', error)
-  throw new DriverAdapterError(convertDriverError(error))
+type PrismaD1Interface<Params> = SqlDriverAdapterFactory & {
+  connectToShadowDb: Params extends D1HTTPParams ? () => Promise<SqlDriverAdapter> : undefined
 }
