+++ packages/adapter-d1/src/index.ts
@@ -1,2 +1,1 @@
-export { PrismaD1AdapterFactory as PrismaD1 } from './d1'
-export { PrismaD1HTTPAdapterFactory as PrismaD1HTTP } from './d1-http'
+export { PrismaD1 } from './d1'
